requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
docopt==0.6.2
filelock==3.9.0
Flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
h5py==3.7.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lightning-utilities==0.5.0
MarkupSafe==2.1.2
matplotlib==3.6.3
multidict==6.0.4
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
Pillow==9.4.0
pipreqs==0.4.11
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.9.0
PyYAML==6.0
requests==2.28.2
scipy==1.10.0
six==1.16.0
timm==0.6.12
torch==1.13.1
torchmetrics==0.11.0
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
yarg==0.1.9
yarl==1.8.2
zipp==3.11.0
